make cogeo-mosaic a default requirements and cleanup
