+++ setup.py
@@ -14,10 +14,10 @@ inst_reqs = [
     "rio-tiler>=2.0b8",
     "rio-cogeo~=2.0a5",
     "rio-tiler-crs==3.0b4",
+    "cogeo-mosaic==3.0a10",
 ]
 extra_reqs = {
     "dev": ["pytest", "pytest-cov", "pytest-asyncio", "pre-commit", "requests"],
-    "mosaic": ["cogeo-mosaic==3.0a10"],
     "server": ["uvicorn", "click==7.0"],
     "lambda": ["mangum>=0.9.0"],
     "deploy": [
