+++ tests/routes/test_mosaic.py
@@ -265,3 +265,13 @@ def test_mosaic_not_found_error(app):
     """
     response = app.get("/mosaicjson", params={"url": "mosaic.json"})
     assert response.status_code == 424
+
+
+def test_validate(app):
+    """test POST /mosaicjson/validate endpoint"""
+    body = read_json_fixture("mosaic.json")
+    response = app.post("/mosaicjson/validate", json=body)
+    assert response.status_code == 200
+
+    response = app.post("/mosaicjson/validate", json={"nope": "oups"})
+    assert response.status_code == 422
