+++ titiler/main.py
@@ -1,25 +1,17 @@
 """titiler app."""
 
-import importlib
-
 import pkg_resources
-from rio_cogeo.cogeo import cog_info as rio_cogeo_info
-from rio_tiler_crs import STACReader
 
 from . import settings, version
 from .db.memcache import CacheLayer
-from .dependencies import PathParams
-from .endpoints import tms
-from .endpoints.factory import TilerFactory
+from .endpoints import cog, mosaic, stac, tms
 from .errors import DEFAULT_STATUS_CODES, add_exception_handlers
-from .models.cog import RioCogeoInfo
 
-from fastapi import Depends, FastAPI, Query
+from fastapi import FastAPI
 
 from starlette.middleware.cors import CORSMiddleware
 from starlette.middleware.gzip import GZipMiddleware
 from starlette.requests import Request
-from starlette.responses import HTMLResponse
 from starlette.templating import Jinja2Templates
 
 template_dir = pkg_resources.resource_filename("titiler", "templates")
@@ -38,68 +30,10 @@ app = FastAPI(
     version=version,
 )
 
-################################################################################
-# COGEO
-cog = TilerFactory(router_prefix="cog")
-
-
-@cog.router.get("/validate", response_model=RioCogeoInfo)
-def cog_validate(
-    src_path: PathParams = Depends(),
-    strict: bool = Query(False, description="Treat warnings as errors"),
-):
-    """Validate a COG"""
-    return rio_cogeo_info(src_path.url, strict=strict)
-
-
-@cog.router.get("/viewer", response_class=HTMLResponse)
-def cog_demo(request: Request):
-    """COG Viewer."""
-    return templates.TemplateResponse(
-        name="cog_index.html",
-        context={
-            "request": request,
-            "tilejson": request.url_for(f"{cog.router_prefix}tilejson"),
-            "metadata": request.url_for(f"{cog.router_prefix}metadata"),
-        },
-        media_type="text/html",
-    )
-
-
 app.include_router(cog.router, prefix="/cog", tags=["Cloud Optimized GeoTIFF"])
-
-################################################################################
-# STAC
-stac = TilerFactory(reader=STACReader, add_asset_deps=True, router_prefix="stac")
-
-
-@stac.router.get("/viewer", response_class=HTMLResponse)
-def stac_demo(request: Request):
-    """STAC Viewer."""
-    return templates.TemplateResponse(
-        name="stac_index.html",
-        context={
-            "request": request,
-            "tilejson": request.url_for(f"{stac.router_prefix}tilejson"),
-            "metadata": request.url_for(f"{stac.router_prefix}info"),
-        },
-        media_type="text/html",
-    )
-
-
 app.include_router(stac.router, prefix="/stac", tags=["SpatioTemporal Asset Catalog"])
-
-
-try:
-    fact = importlib.import_module("titiler.endpoints.factory_mosaic")
-    mosaic = fact.MosaicTilerFactory(router_prefix="mosaicjson")  # type: ignore
-    app.include_router(mosaic.router, prefix="/mosaicjson", tags=["MosaicJSON"])
-except ModuleNotFoundError:
-    pass
-
-
+app.include_router(mosaic.router, prefix="/mosaicjson", tags=["MosaicJSON"])
 app.include_router(tms.router)
-
 add_exception_handlers(app, DEFAULT_STATUS_CODES)
