+++ tox.ini
@@ -2,7 +2,7 @@
 envlist = py36,py37,py38
 
 [testenv]
-extras = test,mosaic
+extras = test
 commands=
     python -m pytest --cov titiler --cov-report term-missing --ignore=venv
 deps=
@@ -51,5 +51,5 @@ ensure_newline_before_comments = True
 line_length = 90
 known_first_party = titiler
 forced_separate = fastapi,starlette
-known_third_party = rasterio,morecantile,stac_tiler,rio_tiler_crs,rio_tiler
+known_third_party = rasterio,morecantile,rio_tiler_crs,rio_tiler,cogeo_mosaic
 default_section = THIRDPARTY
